Fix: resolve comments of pr
