+++ .hintrc
@@ -1,16 +1,13 @@
 {
-  "extends": [
-    "development"
-  ],
-  "browserslist": [
-    "defaults",
-    "not ie 11",
-    "not and_ff <= 110",
-    "not firefox <= 111",
-    "not safari <= 16.3"
-  ],
-  "hints": {
-    "no-inline-styles": "off",
-    "axe/forms": "off"
-  }
+    "extends": ["development"],
+    "browserslist": [
+        "defaults",
+        "not ie 11",
+        "not and_ff <= 110",
+        "not firefox <= 111",
+        "not safari <= 16.3"
+    ],
+    "hints": {
+        "no-inline-styles": "off"
+    }
 }
